Use continuous ramp for loads, concentrations
Previously we used discrete color breaks for loading rates and
concentration rates for catchments and their streams. Now we
switch to a continuous color ramp where each value is inter-
polated within the gradient.

Each gradient is expected to have 5 stops. Beyond this, we
add an addition sixth stop, immediately after the fifth, which
goes from a darker red eventually to black.

This color ramp uses the same colors used for stream coloring
in _variables.scss.