Catch and report GWLF-E errors
Previously errors in GWLF-E would not be caught and reported
to the user, letting them wait up to 260s before failing. This
could lead to a frustrating user experience, as well as extra
burden on our servers as the front-end continued to poll.

This catches and reports errors quickly so the user is alerted
and the servers spared, while also logging to Rollbar allowing
developers to know what happened.