Handle cases with no agricultural streams
When dividing subbasin's agricultural streams with the total
agricultural streams in the AoI to find out the weight factor,
we assumed that the AoI would always have some agricultural
streams. However, as is the case in Carrizozo Creek HUC-10 in
New Mexico, there are some shapes that don't have any such
streams.

To account for this, now we catch the ZeroDivisionError and
default to a weight of 1, which would use the values as is.