Check DRB eligibility against simpler perimeter
Previously we would compare a shape against a high-resolution
GeoJSON of the DRB. This would sometimes fail for shapes that
fell on the edges, causing minor coordinate variations to say
that those shapes were not in the DRB, when in fact they are.

By comparing against a simplified, buffered DRB shape (already
used elsewhere), we ensure that all true shapes are correctly
deemed eligible. Furthermore, since the shape is significantly
simpler, the comparison is much quicker, speedening up general
execution as well.

Refs #2812