Switch boundary details to return Feature
Previously the boundary details endpoint would return a geometry
as GeoJSON, i.e. a MultiPolygon or Polygon. This would not capture
any additional attributes the source data had to offer.

Now we return a GeoJSON Feature, which is treated identically in
the front-end, but comes with additional data packaged with it.
We use the optional properties field of the Feature to add the
HUC id for those tables that support it.

Refs #2814