Add location / coordinate support to search bar
Parses the entered query, and if it matches a coordinate that
falls within the continental United States then shows a "Go
to xxx" result box instead of the suggestions list shown in
other cases. Clicking that box, or pressing Enter while in the
search box, will take the user to the set location, and clear
the search.