Pin Pandas to 0.22.0 so that Ulmo can work
Ulmo does not specify dependency versions, making it vulnerable
to API changes. One such API was pandas.Timestamp().to_datetime,
marked as deprecated since 0.21.0 and finally removed in 0.23.0,
in favor of pandas.Timestamp().to_pydatetime.

Until this fix is in Ulmo, we're relegated to using an older
version of pandas to prevent runtime crashes when querying
CUAHSI data.