Restrict USGS WQP to BiG-CZ sites
In order to minimize potential resource / user experience issues
with the new USGS WQP data source for monitoring data, we limit
access to it to the lesser used BiG-CZ Portal, and hide it from
the Monitor My Watershed which has a larger audience. This will
allow us to identify issues early and fix any found bugs, and
add a little polish before releasing it for the rest.

WQP can be accessed from localhost:8000/?bigcz, portal.bigcz.org,
and staging.portal.bigcz.org.