Always keep streams on top of catchments
Previously, because of the order in which the listeners are added,
when hovering on a catchment the catchment's layer would be sent
to the front, and then the stream. But upon removal of the hover,
the catchment would be sent to back, and then the stream, causing
the stream to be obscured by the catchment's fill.

Instead of sending the streams and catchments to the back, if we
instead bring the parent HUC-12 to the front, the results are more
predictable and manageable. The code for bringing that to the front
(which will now be called more often) is made a bit more efficient
with the use of a simple for loop and short-circuiting the loop
when the right layer is found.