Set full opacity for loads, increase visibility
To better communicate the load concentration values, when a load
is active, we set the catchment opacity to 1.0. At a later stage,
this may be tied to an opacity slider.

To better show the blue lines for streams, which are often atop
a similarly blue section of the basemap, we change the catchments
to be colored White instead of Grey by default. This makes the
blue streams pop a bit more.