Add Subbasin Legend, Opacity Slider
Adds a box to the right of the layer picker with a legend and
an opacity slider for the subbasin related layers. It is shown
when a related layer is selected, and hidden otherwise. When
the opacity slider is adjusted, its value is recorded on the
map model, and made available to all future related layers.

Notably, this is shown on top of the time slider for the climate
layers, with the assumption that these will never be seen at
the same time.