Fix IE11 failure with lingering highlights
In IE11, there are cases when a catchment will stay highlighted
if the mouse moves too fast for `mouseout` to register. In this
case, when the opacity slider is adjusted, the highlighted block
of the `if` expression gets executed rather than the regular one.
Unfortunately, if the mouse isn't hovering on the actual layer,
its `._map` is set to `null`, causing `.bringToFront()` to fail.

To catch this case, we wrap it in a `if (layer._map)` test, to
ensure it only executes when hovering.