Short-circuit failing /values/ requests
In some cases, when the contents of the /details/ request don't
match the variable list in /search/, the start and end date for
the non-matching variables is not computed correctly, instead
stored as an "Invalid date" string. In these cases, the /values/
request will fail, because the string cannot be converted to a
valid date.

When we know we have a valid date, requesting it from /values/
is redundant and expensive. Instead, we now trap it and error
out quickly.