Add support for overlay / combined layers
For cases when a single MMW layer corresponds to two or more
TileLayers, we add the `overlay_codes` layer setting. When set,
this will add the specified layers to a special `overLayers`
group, which will always be atop the base layer. All operations,
such as turning it on or off and setting opacity, will be synced
across the base layer and all the overlayers.