Adjust front-end to poll for results
A HydroShareExportTaskModel is added to poll for results,
set to poll every 6 seconds since this is relatively slow
process. A task model gives us two promises: one for when
the initial request is made, and another for when polling
stops.

We handle errors such as connectivity problems etc in the
start promise failure. We handle response from HydroShare
in the polling promise resolution. The exportToHydroShare
method returns the two promises combined, to resolve when
they are both ready. This isn't currently used, but keeps
the API same as before when we returned the AJAX promise.