Disable HydroShare Autosync
Since now the HydroShare Export can potentially take
quite a while, during which the user can keep making
changes, with Autosync enabled that would simply add
each change to the queue, leading to sitations where
the app would constantly be exporting to HydroShare.

Furthermore, because this export queue isn't visible
to the user, they can navigate away, blocking export
from completion.

Adding a visible queue that is user-manipulatable is
currently out of scope. Instead, our compromise here
is to disable autosyncing completely, and relying on
the user to export manually when they want.