Install numba before dependencies
The next version of GWLF-E requires numba to be present.
Unfortunately, simply specifying it in requirements.txt
does not seem to do the trick, as after installing numba
there is a compilation step, which needs to occur before
GWLF-E is installed, but doesn't occur when installing
via requirements.txt. It does occur when installing
directly via pip.

Thus, we add an explicit pip install numba command via
ansible to execute before the requirements are installed.

Since the post-installation step requires building C++
libraries, we add build-essential to both app and worker
VM. This adds roughly ~30s to the provisioning time.