Indicate changed project for HydroShare export
Since we no longer autosync projects, and the user must
be prompted to export explicitly, we add a message in the
MultiShare Modal under the HydroShare section saying if
the project has changed since last export.

This message is only visible if the date of last export
is less recent than the latest modified_at date of all
the scenarios in the project. Thus, if any scenario is
modified after an export, this message will be shown.

To allow this, we now always update the client scenario
model with the modified_at date from the server.