Add Scenario Item View for GWLFE
This adds a new template and view for GWLFE scenarios
that doesn't include a map. It does include a spot for
showing a scenario color which will correspond with the
charts, to be added in upcoming PRs.

The list of modifications shown on hover isn't quite
right for MapShed. This will be addressed in #2920.