Only export changed scenarios
Previously we always exported all scenarios for
simplicity. Unfortunately, as we bump up against
the limit of how many files can be exported before
a timeout, especially when the record already exists
and each export checks for a file, deletes it, and
then exports it again, being judicious about which
files we send ensures that we almost never time
out.