Add Selection View
This transforms the flat list of Selections into a nested list
of groups and names. Updates the model when selection is changed.
The currently selected item is set to "inactive" silently, and
then the new item is set to "active". The first is done silently
to not fire two events for every change.