Make ChartView more reusable
By extracting TR-55 specific functionality into a
subclass, but leaving common functionality, such
as that of sliding, in the base class.

Also adjust sliding so it works with both types of charts.
In the case of a discrete bar chart, the child of `.nv-group`
is a `g`, not a `rect`. The new selector is more generic and
supports both tags.