Respond to Selection Changes
By moving the model-specific update logic into the model, the
view code can be greatly streamlined. This adds more information
to the selection models, upon selecting which tells the section
model to transform its data for the view.

Also rename `group` to `groupName` and use new `group` for
internal values.