Extract chart helper functions for reuse
We'll need to make a separate render function for the
GWLF-E Water Quality charts, since they need to render
each bar as a separate entity rather than part of the
same series. But there will be some things common to
both the Multi Bar Chart used for TR-55 and the Discrete
Bar Chart used for GWLF-E, so that functionality is
extracted.