Adapt GWLF-E Hydrology Chart to current pattern
Instead of adding and removing the chart a number of times,
simply render it on attach, as is done for all the other
charts.

Also, add a onRenderComplete handler, as done for the other
charts. This is used to slide the bar charts to match the
scenarios, which doesn't really apply here, but we've added
it for uniformity.