Deduplicate vertices to fix line chart issue
Voronoi rendering assumes distinct vertices. When this is
not the case, such as for line charts with multiple series
having identical values (think Current Conditions and an
Unmodified Scenario in MMW), we get errors. This issue
was seen by multiple people (see novus/nvd3#330), and
fixed in novus/nvd3#584. We adopt that fix for our shim
here.

This also corresponds to https://github.com/azavea/nvd3/pull/3