Support downloading text files
Previously, `downloadAsFile` always stringified the
input data, assuming it to be a JSON object. This
was not the case for the TR-55 Compare View download,
which was a CSV string. Stringifying that led to
all the formatting elements, such as quotes and new-
lines, being escaped in the final download, thus
corrupting it.

We replace `downloadAsFile` by two derived methods,
`downloadJson` and `downloadText`, each of which
calls the now renamed `_downloadAsFile` with a
parameter that tells it to stringify or not.

All existing use cases have been replaced by the
appropriate alternative. This allows JSON downloads
to continue to work as before, while also allowing
CSV downloads to work correctly.