Adjust CN for Cover Crops, Conservation Tillage
* Update `makeComputeOutputFn` to pass the dataModel to
  `getOutput`. Since this is the last argument, all present
  implementations will simply ignore it.
* Add `makeCurveAdjustingAgBmpConfig` that takes a `tillFactor`
  and adds a curve number adjustment to the output with a new
  Curve Number value for Cropland.
* Use new config function for Cover Crops, Conservation Tillage
  with respective `tillFactor`s. These values come from #2939
  and #2941.

Since here we need to change the value in an array, not a simple
key in dataModel, we compose a key of the format `{key}__{index}`
to be decoded on the server side. Since both Cover Crops and
Conservation Practices target the Curve Number for Cropland, the
key in both cases is `CN__1`.