Add No Till, Reduced Till MapShed BMPs
While the original card asked for an option button inside
the the Conservation Tillage BMP and change the calculated
values according to the selected option, that turned out
to be very difficult given the current organization of the
code.

Thus, the two new options are added as top-level BMPs. It
is possible that the wording and titles may be changed in
a future PR.