Aggregate GWLF-E Modifications
So far, we've let each GWLF-E BMP supply its own override
dict which replaces values of the baseline GMS. However,
this is no longer possible as we have multiple BMPs that
target the same GMS values, e.g. the three Crop Tillage
BMPs that all set the value of n26, n65, n73, and n81; and
Cover Crops which sets the Curve Number for Cropland CN__1
just as the three Crop Tillage BMPs do.

Thus, we add a method to aggregate GWLF-E modifications,
which using special logic for the above cases, reduces all
active modifications to a single dict of values to override
the baseline GMS values.