Reuse GWLF-E aggregation for front-end
Since we generate a GMS file by sending a fully composed
data model to the back-end, that needs to be updated using
the new aggregation method as well.

The aggregation method is available on the Scenario
model, and passing that to a core/util method would introduce
a cyclic dependency. Thus, the `applyGwlfeModifications`
method is moved to the model as `getModifiedGwlfeGisData`
and the relevant references updated.