Add simple validation
Since GWLF-E modifications interact with each other, it is not
enough to validate them in isolation, but also in aggregate.
This method checks the final override dict for warning signs
and logs them to the console. In the future we may develop a
UI to better communicate to the user.