Guard against CUAHSI service being unavailable
The suds client initialization is moved to as low a scope as can
be done, and the calls to initialize are wrapped in try blocks
which raise appropriate exceptions when running in to availability
issues. In case there are other issues, they will fail with a 500,
but will be limited to those endpoints, and will not bring down
the entire app.

Refs #2972