Make fields more responsive to user input
Previously we would update the field's underlying model value on
a "change" event, which fires when the input loses focus. While
that is more efficient since it updates the model less often, it
is also less responsive, since the user cannot get feedback until
the input has lost focus.

By changing the model update frequency to be on every "keyup", we
allow much faster feedback to the user for when they mistype some-
thing, or as they aim to match a target total value for the land
cover area.

To do this, some of the template had to be changed to allow
parts to be manipulated without a full re-render (which has the
unfortunate side-effect of losing the cursor position in the
textbox). The approach here is similar to React, where the
"props" are kept in the template, but the "state" changes are
done through view methods.