Ensure dropdowns have undo button
Previously, when we listened to the `change` event for
all types of `.form-controls`, we could use one event
listener configuration for input and select. When this
was refactored in 529ba13, since there was never any
`keyup` event fired on the select, changing its value
didn't result in a handler executing.

Furthermore, the UI architecture was very render-oriented,
and didn't play well with the jQuery DOM change style.
After moving a few pieces around in the template and the
styles, we achieve the same visual effect as before, but
now done in a way that allows fast feedback for typing
in a text field.