Add GMS Download button to Scenario Buttons View
This view classically does not correspond to a single scenario,
but this is where they want the GMS download button to be, so
we add it, always querying for the active scenario and then
populating its gis_data in the download form.

Much of the code has been copied from ScenarioDropDownMenuOptionsView,
which still has an Export GMS button. A GMS file can be exported
from both places.

A third place is in ScenarioToolbarView, which is more specific
to the scenario, but that comes after the Scenario Buttons of
Add New Scenario and Compare, and we want this to be seen before,
so it matches to when there's only Current Conditions.