Open Project via Snapshot
First we try to find the project using the resource id, as before.
Then, we try to fetch its project snapshot using a direct request.
Then, we see if the user is logged in and has linked their account
to HydroShare, and if so, try and fetch the project snapshot using
the HS Rest Client. If that also fails, we show an error message
listing the points of failure.