Edit Project via Snapshot
If a user selects "Edit with MMW" in HydroShare, they will be
taken to this new `/project/via/hydroshare/<resource-id>/edit`
endpoint.

If they selected their own project, it will just open.

If they selected someone else's project, it will be copied into
their account.

If they copied someone else's resource on HydroShare and then
edit it from their account, it will be copied to their account
and associated with the copied HydroShare resource.

If they make a copy of their own resource on HydroShare and then
edit it, it will be duplicated in their account and associated
with the new copied HydroShare resource.

In the last two cases, the HydroShare resources will be updated
to point to the newly created projects.

If they are not logged in to MMW, they'll see an error message
prompting them to log in. Dismissing it shows the login modal.