Add Unit Conversion module
There are two top-level keys: METRIC and USCUSTOMARY. Each is
an object containing standardized unit names as keys, and a
value in the shape { name, factor, offset }. The `name` is the
displayed name of the unit. The `factor` is the numerical
conversion value by which a standard value is divided to get
the desired unit. The `offset` is the value which is added to
the converted number to get the desired unit, and will only
be used for temperature.

The `get` function takes a standardized unit name and a value
in the standard unit, and returns an object { value, unit }
which has the converted value and unit name. It reads the user's
preferred unit scheme from the current clientSettings.

For example, to get the area of the current area of interest in
square miles, we can do:

    var areaInM2 = turfArea(App.map.get('areaOfInterest')),
        areaInMiles = units.get('AREA_XL', areaInM2);

    console.log(areaInMiles.value + ' ' + areaInMiles.unit);