Use scheme units for Area, Land, Soil results
This includes the error messages shown to the user when the
area is too large, to the size of the area of interest in the
Analyze stage, to the Land and Soil tabs.

This also updates the tests to use m² instead of m<sup>2</sup>,
to be in line with the new standards.