Always export all files to HydroShare
Previously, since the time needed for exporting grew very
sharply with the number of files, we had employed a number
of tricks to avoid sending more files than necessary.

Thanks to the recent bulk upload work, that is not as
immediate a limit as it used to be. Furthermore, if a user
changes their preferred unit scheme, then all the analyses
and scenarios need to be re-exported, with the new units,
even if they have not numerically changed since the last
export.

To facilitate this, we now export all files every time.