Show correct slider value on slide end
When the slider is dragged, we use two events to update
the numeric value shown next to it: onSliderDragged and
onSliderChanged. The latter is necessary because Internet
Explorer does not support onSliderDragged. That is usually
a no-op, because the value it sets are the same as those
set in onSliderDragged.

Unfortunately, when we updated the code to use Unit Conversions
in https://github.com/WikiWatershed/model-my-watershed/commit/6728d758fbb83dce6f9ff7c13092b60c9df5c7a7#diff-974b9f0d8dedea0247b441534588d17dR573,
we mistakenly started displaying the "under the hood" value,
which is always in inches, rather than the slider's display
value. By showing the slider value, we ensure it is correct
in all unit schemes.