Fix UI crashes for Waste Management BMPs
Because the two Waste Management BMPs are in percent units, they
did not use any of the existing ones and caused a crash in the
front-end.

By adding an explicit unit for PERCENT and making them use that,
we ensure that the crash doesn't happen, and that the configuration
is well specified.