Make toUnit filter more robust
Previously the toUnit filter always expected a unit name,
using which it would transform the input. However, in cases
when the unit was not specified, it would crash the template
rendering.

To make this more robust, now if the unit is unspecified, we
simply return the input as is, effectively acting as a noop.