Try and access uncached UtfGrid tiles
Because of CartoDB/Windshaft#310, UtfGrid tiles from the
server can be incorrect due to internal caching. These
incorrect results can be cached by the CDN and in browsers
for up to 30 days, making it difficult to use the app.

By appending a random querystring to the end of each tile
request for UtfGrid values, we try and bust the cache and
attempt to retrieve fresh values from the server every time,
in the hopes that they will be correct.