Use fallback  if no trace present
The core_job table requires `traceback` to be a non null value,
but if a worker crashes with a segfault or SIGSEGV then no
traceback is available. Saving this error fails, and that
exception cascades through to an unhandled failure.

By providing a fallback value for a traceback when none is
present, we ensure that a failing job is accurately marked
as such, even if there is a segfault and sudden crash of a
worker.