Circumvent GWLF-E SegFaults due to NumPy
A recent release of NumPy introduced a new module _multiarray_umath.
Unfortunately, there are some tools that depend on NumPy and pull
in the latest version as a build dependency. This latest version is
then replaced with the version specified in our requirements.txt,
thus the module is not available at run-time causing the segfaults.

By installing NumPy separately before requirements.txt is gathered,
we ensure that whatever was using the most recent version of NumPy
in its installation now uses the pre-installed one.

We'll have to ensure that the version of NumPy in requirements.txt
matches the one in Ansible.

For details see:

https://github.com/numpy/numpy/issues/11871
https://stackoverflow.com/q/54153886