Separate styles, always specify SQL
Previously, we had a single style file that was used for
every query. The layer name tags in the style were used to pick
what is applied to which layer. Unfortunately, that is no longer
supported, and we must only specify the relevant styles for
each layer in the call. This requires separating the styles into
separate files so they don't interfere with one another.

Also, we can no longer specify a table name and have it generate
the appropriate SQL. The query must be specified explicitly.