Add flag to skip JS tests
As fallout from the Ubuntu Upgrade, testem hangs when running in
CI mode. We have been unable to pinpoint the source of this error,
so to stop CI from hanging every time we add a flag to skip the
tests and configure CI with it.

Inspired by https://github.com/OpenTreeMap/otm-cloud/pull/227