Upgrade Django Rest Framework to 3.9.4
This is the last version that supports Python 2.7.

All serializers need an explicit field list now, so that is
added wherever it was missing. Extraneous serializers that
were not used anywhere were removed.

Read-only complex serialized fields now need explicit code
to save in new instances, so that is also added.