Use request.data instead of request.body
request.body is now a stream instead of a variable, and its
value is read somewhere up the chain, thus it can't be read
again at this point. We switch to request.data instead,
which is already a dict, thus supplanting the json.loads.

Also don't modify request.data, instead make a copy and modify
that, since request.data is now immutable.