Switch to systemctl for the bind mount
A bind mount was introduced for node_modules in 8c5d49a8 to
prevent a race condition. The manner of its implementation
made it apply only immediately after provisioning, and not
when loading a provisioned VM. By moving that mount to
systemctl, we ensure it is always loaded, and ready before
the tiler service starts.