Switch to DRF-YASG for api docs
The Django Rest Swagger version we were previously using was
not compatible with Django 1.11. The most recent version is,
but has a completely different API. Furthermore, it is now
abandoned and no longer actively maintained.

We instead switch to using DRF-YASG, an actively maintained
project which also uses a new API, and supports a newer UI
for Swagger.

It also supports ReDoc, a different API documentation tool,
which we can use in the future if needed.