Document views with OpenAPI schemas
As part of the new style of Swagger document generation,
DRF-YASG uses these OpenAPI schema definitions instead of
the YAML style used previously.

Unfortunately, specifying a request_body in the decorator
marks it as required in the generated documentation. This
is countered via instruction to the contrary in the header
section.