Upgrade Lodash to 4.17.15
This is a big upgrade, unfortuantely necessitated by
security fixes. This upgrade deprecates a number of
Underscore compatible aliases that Lodash supported,
requiring us to rename functions to be more Lodash
compatible.

This migration was largely informed by the instruction
here: https://github.com/lodash/lodash/wiki/Migrating

Also, in cases when both Underscore and Lodash were
used in the same file, we switch to only using Lodash.

If we want to replace Underscore completely and use
Lodash only, we'll have to upgrade Backbone and all
the related libraries. That is a much bigger lift,
and is deferred for a later date.

Also add Lodash as an explicit dependency in the tiler
to upgrade some transitive dependencies.