Add tests for MapShed calculations
Using a sample area of interes and observed output, add
a test that ensures the output matches. This is useful
when making changes to the calculation to ensure that
the numbers do not change significantly when the calcs
are updated.

The dictdiffer package is brought in which allows for
deep comparison of dicts, with a configurable tolerance
for numeric comparison. The tests are currently set to
tolerate differences smaller than 1e-15.

The test is implemented as a management command rather
than a unit test because it needs access to the special
MapShed tables to run the calculations, which are not
available in the Unit Test database. To add the large
tables as fixtures to the Unit Test database would be
too big of a slowdown for the sake of this one test,
and would considerably bloat the repo size. To ensure
the test does not run with the others, it is tagged and
excluded from the test script.