Recalculate Area derived values in back-end
Previously, when the user would override Land Use values,
we would only update the Area list in the MapShed dict.
However, there are a number of other fields in the dict
whose values are derived from the Area list, which were
not updated.

The derivation of those fields is now extracted into its
own method, which is called in the initialization of the
dict, as well as during ingestion of modifications from
the user. If the set of modifications contains updates
to the land use distribution, then the area derived
fields are recalculated.

Notably, the LS calculation is not included in this
refactor. This is because while it does derive its value
from Area, it is also dependent on the geoprocessing
result of `lu_stream_pct`, which isn't saved anywhere.
Because of that, we cannot adjust its value with new
proportions when the user updates land use changes.
Amending this would require a larger refactor, which is
deferred to a later date.

The test introduced in the previous commit serves to
ensure that this refactoring does not alter the
calculations.