Recalculate Area derived values in front-end
The changes made in the previous commit recalculate the
derived values for running the MapShed model. The changes
in this commit recalculate them for the data model in the
front-end, which allows for other modifications to pick up
on the land use changes.

For example, if the user changes the area of Cropland in
the Land Cover modal, the Area of Row Crops should change
correspondingly for the respective Conservation Practices.
This will now happen, as `n23` (the model field representing
the Area of Row Crops), amongst others, is now updated.