Isolate failing code and add failing test
The bit of code that generates a 1 square km box from a
given point is isolated into a utility function so it can
be tested separately. Also add a test that exercises the
code with sample values taken from production (where the
failure does not occur).

The failure is likely due to a subtle change in Lodash
behavior caused by the upgrade in 6ac8212a.