Unroll and simplify swNe calculation
This was previously done by a dense sequence of Lodash
methods, some of which were affected by the Lodash upgrade
and had stopped working correctly.

Now we make the final array by hand using direct values,
foregoing any Lodash functionality. This makes the code
easier to read and understand, and the tests pass.