Hide subbasin legend when not in catchment mode
Previously the subbasin legend would remain active
outside the catchment view, which was the only view
representing the information that legend corresponded
to. In some cases, it would stay on even after exiting
the subbasin view entirely.

Now it is removed from view whenever the subbasin
catchment mode is inactive.