Upgrade Testem
The old version was not compatible with new node.

The latest version of Testem uses the configuration JSON to
set the browsers rather than command line arguments, so we
move "Firefox" from test.sh and testem.json

The `beforeEach` behavior was subtly changed so it only runs
every time for `it`s at the same level, but if they are nested
within `describe`s it only runs for each `describe`, so some
were nested out one level to restore the `beforeEach` behavior.

There were some subtle numerical calculation changes to some
tests under `turfArea`, even though `turfArea` itself was not
upgraded, but they are upgraded to match the new numbers.

Testem also removed embeded frameworks in v1.0.0, so we now
copy over the `mocha` files from node_modules to the static
directory so the custom test page has access to it. For details
see:

  - https://github.com/testem/testem/releases/tag/v1.0.0
  - https://github.com/testem/testem/pull/706