Scaffold worksheet endpoint
This will be available at /api/modeling/worksheet/,
the first modeling method available in the public API.
This is done for future compatibility, where we expect
all modeling methods to migrate to eventually. Also
because the input this takes (just a shape) is closer
to an Analyze method than the existing Modeling methods.