Add Spreadsheet Tool template
This is also available in https://github.com/WikiWatershed/MMW-BMP-spreadsheet-tool/,
although the version there is older than the one we are
supposed to use, so the newer version is added directly
to the repo.

From a runtime perspective, it is better to have the template
available immediately on disk for dynamic generation, rather
than downloading it from GitHub every time. This protects us
from GitHub outages.

This does cause some redundancy, and the need to keep the two
versions in sync. An alternative to this direct inclusion is
to include the source repo as a pinned submodule, and update
the submodule SHA reference when that repo is updated.

However, given that the original repo has an older version
of the file, I'm including the new version here for now. The
strategy for maintenance going forward is a policy decision
that should be made by the team.