Add export endpoint for generating the worksheet
The /api/modeling/worksheet endpoint will be used to
generate a JSON blob of values, which will then be
submitted to this endpoint to generate the actual
spreadsheets.

The expected input is still to be clarified, but will
be something along the lines of this:

[
	{
		"huc12": { ... },
		"aoi": { ... },
	},
	{
		"huc12": { ... },
		"aoi": { ... },
	},
]

For every pair of values, we copy the worksheet template
and fill in the values, save the new files in a temporary
directory, zip them all up in memory, delete the directory,
and return the zip file.