Add Active River Area numbers to Land analysis
We are incorporating Active River Area analysis into the Land
tab. Currently, this is tightly coupled with the endpoint in
the back-end, so we incorporate it in to the NLCD analyze end-
point.

The input values do not change, but the output of the endpoint
now includes an `active_river_area` key which has the area of
land under each land use which qualifies as an active river
area.

This uses the same Active River Area raster which is used to
visualize the layer in the layer picker.

Since now the same tab is doing both calculations, we switch
to using a double RasterGroupedCount operation, which pairs
the NLCD results with the Active River Area results. To get
just the NLCD values as before, we sum them up ignoring the
Active River Area values. The Active River Area values will
be summed up on the front-end.