Switch from NOWKAOI to NOCACHE
Previously, we used NOWKAOI as a marker in place of an actual
WKAOI for shape identifier in a multi-operation geoprocessing
request body to indicate that the result should not be cached.
This was accpetable since the only time when multiple shapes
were specified was in the case of sub-HUCs of a larger HUC, and
they would always have WKAOIs. When a custom shape was given,
it would be a single shape, and thus did not need to be uniquely
identified.

However, now we will use multi-operation geoprocessing to run
calculations on possibly a number of anonymous areas of interest,
such as when an area of interest crosses HUC-12 boundaries and
must be clipped to each HUC-12 to make a pair with it. In this
case we need to identify the shapes, but not cache them as they
are not well-known and are incredible unlikely to repeat.

Thus, we switch from using a NOWKAOI identified to a NOCACHE
prefix. When the NOCACHE prefix is seen, caching is skipped.
The suffix of the identified is the name of the HUC-12 to which
the shape belongs, which allows for identification and extraction
at a later time.