Surface n_count in nlcd_soil processing
Since for the worksheet export, we need both the MapShed
results (which include land classification in its terminology)
and the NLCD classification, we surface it in the nlcd_soil
processing step. This will be used to store it for use later
in the workflow.

Adding this additional field does not disrupt any existing code,
because these fields are referenced by name not index, thus it
will only be used when explicitly called, and will not affect
the access of any other field.