Implement PA DEP workflow endpoint
Finalize the working of the PA DEP workflow enpoint. This takes
an area of interest, matches it to HUC-12s, collects all the
AoIs and HUC-12s together, and then runs them through a Celery
workflow that looks like this:

     geoprocessing               geoprocessing
           |                           |
 collect_worksheet_aois     collect_worksheet_wkaois
           |                           |
           +-------------+-------------+
                         |
                  collect_worksheet
                         |
                 transform_worksheet

Using the tasks from the previous commit. The results are saved
as JSON, available for fetching by the front-end. They can be
POSTed as is to the /export/worksheet endpoint to get a ZIP
file of the Excel worksheets and GeoJSONs of the clipped AoIs
in this workflow.