Refactor to use a linear Celery workflow
Using non-linear Celery workflows led to cases where a silent
error would trigger infinite chord_unlock loops which would
saturate the Celery worker. This is now avoided by using a
linear workflow which catches and accounts for errors much
better.

Furthemore, by offloading all the work to the Celery task,
even database interaction, we get a really fast response for
the initial submission request. The Celery task is decorated
with a longer timeout, which allows it to run for a while.
Polling on the front-end will work identically as in other
cases.