Move worksheet transformation logic to export
Previously the calculation logic was coupled with the
format transformation logic. This was unnecessary, and
was hiding calculations that could be useful in and of
themselves as a seperate endpoint.

Now this is paired with the export endpoint. If, in the
future, we need to specify different kinds of Excel
exports, we can choose which kind of transformation
will be used. Currently, the PA DEP transformation is
the default.