Add front-end button for generating worksheet
A button is added to the Model tab of the Analyze section.
Visually, there is a dividing line between this new button and
the two existing modeling buttons.

When the button is clicked, we send a request to the initial
data gathering endpoint and poll for results. A "Gathering data"
message is shown beneath the button with a spinner. The button
itself is disabled, so it cannot be double clicked.

When the polling finishes successfully, a "Download worksheet"
message is shown beneath the button with a check mark, which is
clickable. Clicking this submits a hidden form which had been
populated with the results of the first endpoint, which opens
a new tab and downloads the file.

In case of error gathering data, an error message is shown.

This process is divided into these two steps on the front-end
because if we open a new tab or window programmatically as the
effect of a non-user action (such as the first promise finishing),
browsers treat that as a pop-up and block it by default. Rather
than having users deal with that and manually add an exception
for MMW, we have the user explicitly click some text to start
the download process.