Add ARA Perimeter to test if analysis should work
The ARA Raster does not have data outside the North
East and Mid Atlantic regions, and when requests for
outside that area go through they cause exceptions
in the geoprocessing service.

This perimieter, drawn by hand using GeoJSON.io, is
a rough approximation of the outline of the raster.
It errs on the side of conservatism, i.e. it is smaller
than the true perimeter.